analytics-service/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
pymongo==3.12.0
werkzeug==2.0.3